+++ app/src/main/java/com/example/ceilingmeasurer/data/fake/FakeLocalRepoImpl.kt
@@ -1,0 +1,4 @@
+package com.example.ceilingmeasurer.data.fake
+
+class FakeLocalRepoImpl {
+}
