+++ app/src/main/java/com/example/ceilingmeasurer/data/room/dao/CeilingDAO.kt
@@ -1,7 +1,7 @@
-package com.example.ceilingmeasurer.model
+package com.example.ceilingmeasurer.data.room.dao
 
 import androidx.room.*
-import com.example.ceilingmeasurer.model.Tables.CeilingEntity
+import com.example.ceilingmeasurer.data.room.tables.CeilingEntity
 
 @Dao
 interface CeilingDAO {
