+++ app/src/main/java/com/example/ceilingmeasurer/domain/LocalRepo.kt
@@ -1,0 +1,11 @@
+package com.example.ceilingmeasurer.domain
+
+import com.example.ceilingmeasurer.data.room.tables.OrderEntity
+
+interface LocalRepo {
+
+    fun getAllOrders(): List<OrderEntity>
+
+    fun saveOrderEntity(orderEntity: OrderEntity)
+
+}
