--- app/src/main/java/com/example/ceilingmeasurer/model/LocalRepository.kt
@@ -1,11 +1,0 @@
-package com.example.ceilingmeasurer.model
-
-import com.example.ceilingmeasurer.model.Tables.OrderEntity
-
-interface LocalRepository {
-
-    fun getAllOrders(): List<OrderEntity>
-
-    fun saveOrderEntity(orderEntity: OrderEntity)
-
-}
