+++ app/src/main/java/com/example/ceilingmeasurer/ui/ClientsFragment.kt
@@ -1,4 +1,4 @@
-package com.example.ceilingmeasurer
+package com.example.ceilingmeasurer.ui
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
