+++ app/src/main/res/layout/fragment_clients.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:gravity="center"
-    tools:context=".ClientsFragment">
+    tools:context=".ui.ClientsFragment">
 
     <TextView
         android:layout_width="wrap_content"
